Desenvolvido camada de repositório para a entidade User
