+++ src/main/kotlin/com/icat/orboarding/user/adapters/outbound/persistence/mysql/entities/RestaurantEntity.kt
@@ -1,21 +1,21 @@
-package com.icat.orboarding.user.adapters.outbound.persistence.entities
+package com.icat.orboarding.user.adapters.outbound.persistence.mysql.entities
 
 import org.hibernate.annotations.UpdateTimestamp
 import java.time.LocalDateTime
 import javax.persistence.*
 
 @Entity
-@Table(name = "delivery")
-class DeliveryEntity(
+@Table(name = "restaurant")
+class RestaurantEntity(
     @Id
     @Column(unique = true, nullable = false, length = 50)
     val id: String,
 
-    @Column(name = "full_name", nullable = false, length = 100)
-    val fullName: String,
+    @Column(nullable = false, length = 100)
+    val name: String,
 
-    @Column(unique = true, nullable = false, length = 11)
-    val cpf: String,
+    @Column(unique = true, nullable = false, length = 14)
+    val cnpj: String,
 
     @Column(unique = true, name = "image_url", nullable = false)
     val imageUrl: String,
