+++ src/main/kotlin/com/icat/orboarding/user/adapters/outbound/persistence/mysql/entities/UserEntity.kt
@@ -1,7 +1,9 @@
-package com.icat.orboarding.user.adapters.outbound.persistence.entities
+package com.icat.orboarding.user.adapters.outbound.persistence.mysql.entities
 
+import org.hibernate.annotations.CreationTimestamp
 import org.hibernate.annotations.UpdateTimestamp
 import java.time.LocalDateTime
+import java.util.UUID
 import javax.persistence.*
 
 @Entity
@@ -9,7 +11,7 @@ import javax.persistence.*
 class UserEntity(
     @Id
     @Column(unique = true, nullable = false, length = 50)
-    val id: String,
+    val id: String? = UUID.randomUUID().toString(),
 
     @Column(unique = true, nullable = false, length = 100)
     val email: String,
@@ -26,6 +28,7 @@ class UserEntity(
     @OneToOne(mappedBy = "userEntity")
     val deliveryEntity: DeliveryEntity? = null,
 
+    @CreationTimestamp
     @Column(name = "created_at", updatable = false)
     val createdAt: LocalDateTime? = null,
