+++ src/main/kotlin/com/icat/orboarding/user/adapters/outbound/persistence/mysql/repositories/UserRepository.kt
@@ -1,0 +1,10 @@
+package com.icat.orboarding.user.adapters.outbound.persistence.mysql.repositories
+
+import com.icat.orboarding.user.adapters.outbound.persistence.mysql.entities.UserEntity
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+
+@Repository
+interface UserRepository : JpaRepository<UserEntity, String> {
+    fun existsByEmailContainsIgnoreCase(email: String): Boolean
+}
