+++ src/main/kotlin/com/icat/orboarding/user/application/domain/UserDomain.kt
@@ -1,0 +1,4 @@
+package com.icat.orboarding.user.application.domain
+
+class UserDomain {
+}
