+++ src/main/kotlin/com/icat/orboarding/user/application/ports/inbound/UserServicePort.kt
@@ -1,0 +1,4 @@
+package com.icat.orboarding.user.application.ports.inbound
+
+interface UserServicePort {
+}
