+++ src/main/kotlin/com/icat/orboarding/user/application/ports/outbound/UserPersistencePort.kt
@@ -1,0 +1,8 @@
+package com.icat.orboarding.user.application.ports.outbound
+
+import com.icat.orboarding.user.application.domain.UserDomain
+
+interface UserPersistencePort {
+    fun emailAlreadyRegistered(email: String): Boolean
+    fun createUser(userDomain: UserDomain): UserDomain
+}
