+++ src/main/kotlin/com/icat/orboarding/user/application/services/UserService.kt
@@ -1,0 +1,3 @@
+package com.icat.orboarding.user.application.services
+
+class UserService
